requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
tensorflow==2.13.0
matplotlib==3.7.2
seaborn==0.12.2
yfinance==0.2.18
newsapi-python==0.2.6
fredapi==0.5.1
ta==0.10.2
plotly==5.15.0
requests==2.31.0
python-dotenv==1.0.0
jupyter==1.0.0
tqdm==4.65.0
textblob==0.17.1
pytz==2023.3
streamlit==1.24.0
